so it doesnt appear on all hearts
